Kill sendThread after recvThread in Network
Killing sendThread closes the socket. Because of the way it was handeld
until now, it sometimes happened, that the recvThread tried to read from
the socket after the sendThread had already closed it. This was
problematic as it caused an exception, which was not caught and
therefore printed to stdout.